Move Canvas and Show Diff to View menu
Canvas and Show Diff are view-related actions, not file operations.
Move them from the File menu (CommandGroup replacing .newItem) to
the View menu (CommandGroup replacing .sidebar), alongside the
existing Toggle Left Sidebar item.
