+++ supacode/Commands/WorktreeCommands.swift
@@ -96,14 +96,6 @@ struct WorktreeCommands: Commands {
       )
       .help("New Worktree (\(AppShortcuts.newWorktree.display))")
       .disabled(!repositories.canCreateWorktree)
-      Button("Canvas") {
-        store.send(.repositories(.toggleCanvas))
-      }
-      .keyboardShortcut(
-        AppShortcuts.toggleCanvas.keyEquivalent,
-        modifiers: AppShortcuts.toggleCanvas.modifiers
-      )
-      .help("Canvas (\(AppShortcuts.toggleCanvas.display))")
       Button("Archived Worktrees") {
         store.send(.repositories(.selectArchivedWorktrees))
       }
@@ -129,22 +121,6 @@ struct WorktreeCommands: Commands {
       .keyboardShortcut(.return, modifiers: .command)
       .help("Confirm Worktree Action (⌘↩)")
       .disabled(confirmWorktreeAction == nil)
-      Button("Show Diff") {
-        let repos = store.repositories
-        guard let worktreeID = repos.selectedWorktreeID,
-          let worktree = repos.worktree(for: worktreeID)
-        else { return }
-        DiffWindowManager.shared.show(
-          worktreeURL: worktree.workingDirectory,
-          branchName: worktree.name,
-        )
-      }
-      .keyboardShortcut(
-        AppShortcuts.showDiff.keyEquivalent,
-        modifiers: AppShortcuts.showDiff.modifiers
-      )
-      .help("Show Diff (\(AppShortcuts.showDiff.display))")
-      .disabled(!hasActiveWorktree)
       Button("Refresh Worktrees") {
         store.send(.repositories(.refreshWorktrees))
       }
